+++ package.json
@@ -23,7 +23,7 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
-    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-select": "^4.2.1",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
