+++ src/App.js
@@ -1,37 +1,19 @@
-import React, { useEffect, useState } from 'react';
-import Header from './components/Header';
-import MovieTable from './components/MovieTable';
-import Footer from './components/Footer';
-import NavBar from './components/NavBar';
-import styles from './assets/styles/main.scss';
-import MovieContext from './MovieContext';
+import React from 'react';
+import {
+  Route, BrowserRouter, Switch,
+} from 'react-router-dom';
+import Home from './pages/Home';
+import Page404 from './pages/Page404';
 
-export const App = () => {
-  const [selectedMovie, setSelectedMovie] = useState(null);
-
-  const resetSelectedMovie = () => {
-    setSelectedMovie(null);
-  };
-
-  useEffect(() => {
-    window.scrollTo(0, 0);
-  }, [selectedMovie]);
-
-  return (
-    <div className={styles.mainContainer}>
-      <div className={`${styles.headerContainer} ${styles.container}`}>
-        <Header
-          selectedMovie={selectedMovie}
-          onCloseMovieDetails={resetSelectedMovie}
-        />
-      </div>
-      <div className={styles.container}>
-        <NavBar />
-        <MovieContext.Provider value={{ selectedMovie, setSelectedMovie }}>
-          <MovieTable />
-        </MovieContext.Provider>
-        <Footer />
-      </div>
-    </div>
-  );
-};
+export const App = () => (
+  <BrowserRouter>
+    <Switch>
+      <Route path="/" exact>
+        <Home />
+      </Route>
+      <Route>
+        <Page404 />
+      </Route>
+    </Switch>
+  </BrowserRouter>
+);
