+++ src/pages/Home/Home.module.scss
@@ -1,0 +1,1 @@
+.Home {}
