+++ src/pages/Page404/Page404.module.scss
@@ -1,0 +1,7 @@
+.Page404 {
+    color: white;
+
+    h1 {
+        font-size: 10em;
+    }
+}
