+++ webpack.dev.js
@@ -8,5 +8,6 @@ module.exports = merge(common, {
     hot: true,
     host: 'localhost',
     port: 8081,
+    historyApiFallback: true,
   },
 });
